added ability to track routing from one point in the app, gonna be used to move camera object
