+++ src/routing/routes.js
@@ -1,0 +1,9 @@
+export const routes = {
+    home: "/",
+    about: "/about",
+    contact: "/contact",
+};
+
+export const transitionObjects = Object.entries(routes).map(([key, value]) => {
+    return { to: value };
+});
