+++ src/routing/views.js
@@ -1,24 +1,27 @@
-import React from "react";
+import React, { useEffect } from "react";
 
 import { Route, Routes } from "react-router-dom";
 import { useView } from "./view-context";
-import { CameraLandingPage } from "../landing-page";
 import { AboutPage } from "../pages/about/about-page";
 
 import { OrbitImages } from "../landing-page/components/orbit-images";
 import useImageState from "../landing-page/state/image-state";
 import { CameraView } from "./camera-view";
+import { routes } from "./routes";
 
 export function Views() {
     const { path } = useView();
 
+    useEffect(() => {
+        console.log({ path });
+    }, [path]);
+
     return (
         <Routes location={path}>
-            <Route path="/" element={<CameraLandingPage />} />
-            {/* <Route path="/about" element={} /> */}
-            <Route path="/about" element={<AboutPage />} />
-            <Route path="/camera3" element={<TestView3 />} />
-            <Route path="/camera4" element={<TestView4 />} />
+            <Route path={routes.home} element={<LandingPage />} />
+            <Route path={routes.about} element={<AboutPage />} />
+            <Route path={routes.contact} element={<TestView3 />} />
+            {/* <Route path="/camera4" element={<LandingPage />} /> */}
             {/* <Route path="/landing-page" element={<TestView />} /> */}
             {/* <Route path="" element={<LandingPageScene />} /> */}
             {/* <Route path="/test" element={<TestView />} /> */}
@@ -31,7 +34,7 @@ const TestView3 = () => {
     return <CameraView />;
 };
 
-const TestView4 = () => {
+const LandingPage = () => {
     const images = useImageState((state) => state.images);
     const tapCamera = useImageState((state) => state.tapCamera);
